Relocate non-routes out of /routes files, into utils.cjs
Move all the non-route files

Update all module exports

Import utils.cjs as util in events.cjs, user.cjs, test_api.js

Make sure all calls in aforementioned files to these functions leads with util.
